version: add basic lk version handling
lk does not have any versioning support except of the one hard coded in
the version variable of the fastboot implementation.
We need some versioning support for lk builds and development.
Add a basic versioning support through the makefile.
This makes it possible to define statically the s/w version.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>