msm8960: add LEDs enablement code
cm-qs600 is used on sb-qs600 base board and forms sbc-qs600 as an
evaluation board set. It is also used inside the Utilite2 computer.
Both sbc-qs600 and Utilite2 have a kind of power on LED which should
be enabled during the early boot stage (but not too early) and indicate
that the device has booted to some extent.
Enable the LED during the early boot of LK boot loader.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>